+++ src/lib/output/router.ts
@@ -7,7 +7,7 @@ import {
     type Reflection,
     ReflectionKind,
 } from "../models/index.js";
-import { createNormalizedUrl } from "#utils";
+import { createNormalizedUrl } from "#node-utils";
 import { Option, type TypeDocOptionMap } from "../utils/index.js";
 import { Slugger } from "./themes/default/Slugger.js";
 import { getHierarchyRoots } from "./themes/lib.js";