+++ src/lib/serialization/components.ts
@@ -4,7 +4,7 @@ import type { ModelToObject } from "./schema.js";
 /**
  * Represents Serializer plugin component.
  *
- * Like {@link Converter} plugins each {@link Serializer} plugin defines a predicate that instructs if an
+ * Like Converter plugins each {@link Serializer} plugin defines a predicate that instructs if an
  * object can be serialized by it. This is done dynamically at runtime via a `supports` method.
  */
 export interface SerializerComponent<T extends object> {