+++ src/lib/utils-common/index.ts
@@ -1,5 +1,5 @@
 // utils-common includes utilities which don't depend on anything, and thus can be used
-// within the models export which is suitable for bundling.
+// within the browser-utils export which is suitable for bundling.
 
 export * from "./array.js";
 export * from "./declarationReference.js";
@@ -7,7 +7,6 @@ export * from "./enum.js";
 export * from "./events.js";
 export * from "./general.js";
 export * from "./hooks.js";
-export * from "./html.js";
 export * from "./i18n.js";
 export * from "./index.js";
 export * as JSX from "./jsx.js";